requirements.txt
fastapi
uvicorn
gitpython
jinja2
python-multipart
